Refactor: NavigationGraph.kt and BottomNavItem.class added from MainActivity.kt
